abbr & blockquote transformers, pygments cache
Changed the regex-based abbreviations filter to instead be a pandoc AST
transformer, working by a regex-based abbreviation collector and then
fed into the actual abbreviation replacer transformer. This affords some
context on the document structure, avoiding the issue where abbr tags
would be inserted in codeblocks, headers, etc. As of now, the
transformer only operates on paragraphs.

Added a blockquote transformer to surround each blockquote with hr tags
to avoid having to do that in javascript. Consequently changed css
styling from separate classes to :first-child/:last-child pseudoclasses.

Implemented a pygments caching system where the code's SHA1 hash is
found and if a file with that name is found in the cache store it is
read to retrieve the pygments-highlighted code. Otherwise, a separate
pygments process is spawned and piped into as was already the case. This
speeds things up a bit.